+++ react-components/src/bundle.js
@@ -1,6 +1,6 @@
 import '../../core/sass/main.scss'
 
-import SignupModal from '@src/views/SignupModal/Modal'
+import SignupModal from '@src/views/SignupModal/Base'
 import IndustriesModal from '@src/views/IndustriesModal/Base'
 import CountriesModal from '@src/views/CountriesModal/Base'
 import LoginModal from '@src/views/LoginModal/Modal'
@@ -15,13 +15,13 @@ import ComponentInitialiser from './initialisers/common.initialiser'
 new ComponentInitialiser();
 
 export default {
-  setConfig: Services.setConfig,
-  SignupModal,
-  LoginModal,
-  Tour,
+  CountriesModal,
   createSectorChooser,
   IndustriesModal,
-  CountriesModal,
+  LoginModal,
+  setConfig: Services.setConfig,
+  SignupModal,
   STEP_CREDENTIALS,
   STEP_VERIFICATION_CODE,
+  Tour,
 };